requirements.txt
Mezzanine==4.0.1
dj-database-url==0.3.0
python-decouple==3.0
gunicorn==19.4.5
psycopg2==2.6.1
django-extensions==1.6.1
django-test-without-migrations==0.4